Initial implementation of MetadataManager
